created a demo, added one enemy, the player, ability to move, one steal ability
created a demo? of what the game could look like. I added some really rough sprites to fit the theme. Right now, you can shoot with left click but it is instant and there is no visual cue. As well as this, if you press space while facing an enemy, it kills the enemy while changing an aspect of the player (steal) The only one that is in there right now is to increase the speed by 2 times.  Next to implement is a functioning death state and implement more steals, and add visual cues to the shooting