Added Bullet Tester Scene
In order to see how bullets could work in the game, I created a bullet tester scene that uses a premade bullet scene to spawn in bullets. The Basic Bullet is scripted to move right automatically and (currently bugged) destroy itself on contact with something else.
The Bullet Tester scene is a debug scene with a dummy and several UI buttons that allow you to change options for a multishot pattern and to increase the firing speed.