added CopyBullet, implemented a bullet control system and attached it to player, created new test scene to demonstrate copying
Created a "Bullet Control" scene that handles firing of bullets (basically a retooling of the bullet tester scene from a while back). Press or hold Space to fire your normal bullet, and press B to fire the copy bullet. Attached the bullet control module to the player scene screated by Jordan previously. Created the "Copy Tester" scene which demonstrates copying; The top "enemy" on the right holds the Rapid Fire ability, the middle has the Wide Beam ability, and the bottom has the Piercing ability.