Changed copy bullet shooting; updated Copy Tester to demonstrate
The copy bullet shooting can no longer be shot all the time, and it must be charged before it can fire (in the future, charging will be done by killing enemies). Copy Tester now has a progress bar and a button that charges the copy bullet. Also changed it so that when both the "shoot" and "copy" controls are pressed down, the copy bullet has priority over the normal shot.